Restructure of python converters
