+++ gldcore/rt/Makefile.mk
@@ -43,3 +43,5 @@ dist_pkgdata_DATA += gldcore/rt/triplex_meter_b.png
 dist_pkgdata_DATA += gldcore/rt/triplex_meter_g.png
 dist_pkgdata_DATA += gldcore/rt/triplex_meter_k.png
 dist_pkgdata_DATA += gldcore/rt/triplex_meter_r.png
+dist_pkgdata_DATA += gldcore/rt/json2glm.py
+dist_pkgdata_DATA += gldcore/rt/json2txt.py
