+++ python/Makefile.mk
@@ -2,7 +2,6 @@ dist_pkgdata_DATA += python/csv_merge/csv_merge.py
 dist_pkgdata_DATA += python/example/house.glm
 dist_pkgdata_DATA += python/example/tstat_commit.py
 dist_pkgdata_DATA += python/example/tstat_init.py
-dist_pkgdata_DATA += python/jsonglm/json2glm.py
-dist_pkgdata_DATA += python/jsonglm/test_json2glm.py
+dist_pkgdata_DATA += python/json2glm/test_json2glm.py
 dist_pkgdata_DATA += python/volt_dump/meter_record.py
 dist_pkgdata_DATA += python/volt_dump/voltdump.py
